MFC: Fix by cloder@
Fix a potential infinite loop that could be triggered by an invalid BGP
packet.  Partially from tcpdump.org, with additional paranoia added so
that the problem is actually fixed.

+++ sys/dev/rcons/raster_text.c
@@ -1,4 +1,5 @@
-/*	$NetBSD: raster_text.c,v 1.2 1995/10/04 23:57:22 pk Exp $ */
+/*	$OpenBSD: raster_text.c,v 1.4 2002/03/14 03:16:07 millert Exp $ */
+/*	$NetBSD: raster_text.c,v 1.3 1995/11/24 23:50:56 cgd Exp $ */
 
 /*-
  * Copyright (c) 1991, 1993
@@ -16,11 +17,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *	This product includes software developed by the University of
- *	California, Berkeley and its contributors.
- * 4. Neither the name of the University nor the names of its contributors
+ * 3. Neither the name of the University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
  *
@@ -45,6 +42,7 @@
 
 #ifdef _KERNEL
 #include <sys/param.h>
+#include <sys/systm.h>
 #include <dev/rcons/raster.h>
 #ifdef COLORFONT_CACHE
 #include <sys/malloc.h>
@@ -52,13 +50,13 @@
 #endif
 #else
 #include <sys/types.h>
-#include <dev/rcons/raster.h>
+#include <string.h>
+#include "raster.h"
 #ifdef COLORFONT_CACHE
 #include <malloc.h>
 #define NEW(size) malloc(size)
 #endif
 #endif
-
 
 /* Draws text.  Returns 0 on success, -1 on failure. */
 int
@@ -67,7 +65,7 @@ raster_text( r, x, y, rop, rf, text )
     int x, y;
     int rop;
     struct raster_font* rf;
-    unsigned char* text;
+    unsigned char *text;
     {
     return raster_textn( r, x, y, rop, rf, text, strlen( text ) );
     }
@@ -79,12 +77,12 @@ raster_textn( r, x, y, rop, rf, text, n )
     int x, y;
     int rop;
     struct raster_font* rf;
-    unsigned char* text;
+    unsigned char *text;
     int n;
     {
     int clip;
     int x1, y1;
-    struct raster_char* c;
+    struct raster_char *c;
     struct raster* charrast;
     int i;
     register unsigned char ch;
@@ -234,7 +232,7 @@ raster_alloc( width, height, depth )
 	return (struct raster*) 0;
     linelongs = ( ( width * depth + 31 ) >> 5 );
     r = (struct raster*)
-	NEW( sizeof(struct raster) + height * linelongs * sizeof(u_long));
+	NEW( sizeof(struct raster) + height * linelongs * sizeof(u_int32_t));
     if ( r == (struct raster*) 0 )
 	return (struct raster*) 0;
 
@@ -242,7 +240,7 @@ raster_alloc( width, height, depth )
     r->height = height;
     r->depth = depth;
     r->linelongs = linelongs;
-    r->pixels = (u_long*) (r + 1);
+    r->pixels = (u_int32_t*) (r + 1);
     r->data = (caddr_t) 0;
     return r;
     }
